+++ keyboards/vinhcatba/uncertainty/rules.mk
@@ -1,4 +1,4 @@
-
+SRC += bongo.c
 # using external i2c eeprom
 EEPROM_DRIVER = i2c
